requirements.txt
instaloader>=4.8
google-generativeai>=0.8.5
pandas>=2.0
python-dotenv>=1.0
python-dateutil>=2.8.2
cryptography>=40.0.0
pytest>=7.0
streamlit>=1.28.0
plotly>=5.17.0
